requirements.txt
Django==3.0.1
psycopg2-binary==2.8.4
djangorestframework==3.11.0
djangorestframework-gis=0.15
